chore: move more github code into github module
And make paths relative to .github/workflows/
