+++ cargo-dist-schema/src/lib.rs
@@ -159,65 +159,6 @@ pub struct GithubCiInfo {
     pub external_repo_commit: Option<String>,
 }
 
-/// A github action workflow step
-#[derive(Debug, Clone, Serialize, Default, Deserialize, JsonSchema)]
-#[serde(rename_all = "kebab-case")]
-pub struct GithubJobStep {
-    /// A step's ID for looking up any outputs in a later step
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-
-    /// If this step should run
-    #[serde(default)]
-    #[serde(rename = "if")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub if_expr: Option<serde_json::Value>,
-
-    /// The name of this step
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-
-    /// The identifier for a marketplace action or relative path for a repo hosted action
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uses: Option<String>,
-
-    /// A script to run
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub run: Option<String>,
-
-    /// The working directory this action sould run
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub working_directory: Option<String>,
-
-    /// The shell name to run the `run` property in e.g. bash or powershell
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub shell: Option<String>,
-
-    /// A map of action arguments
-    #[serde(default)]
-    pub with: BTreeMap<String, serde_json::Value>,
-
-    /// Environment variables for this step
-    #[serde(default)]
-    pub env: BTreeMap<String, serde_json::Value>,
-
-    /// If this job should continue if this step errors
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub continue_on_error: Option<serde_json::Value>,
-
-    /// Maximum number of minutes this step should take
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timeout_minutes: Option<serde_json::Value>,
-}
-
 /// Github CI Matrix
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
 pub struct GithubMatrix {
